+++ src/index.ts
@@ -1,2 +1,12 @@
-export { get, getter, Options, Path }          from './getter'
-export { default as parser, default as split } from './parser'
+export {
+    Options,
+    Path,
+    get,
+    getter,
+} from './getter'
+
+export {
+    default as parse,
+    default as parser,
+    default as split
+} from './parser'